backend/requirements.txt
fastapi
uvicorn
langgraph
langchain-openai
mcp
pydantic
python-dotenv
langchain-core
